Do not model has_sig property
For static non-malleability guarantees, there's "s,e,f" type modifiers.
has_sig property is a runtime property. Modelling it might be viable,
but it increases complexity of the model (and the time to check it),
and requires much better understanding of the satisfaction algorithm
to model it correctly, than what I have now. Therefore I decided to
drop it altogether for now.